+++ path_login.go
@@ -136,6 +136,19 @@ func (b *azureAuthBackend) pathLogin(ctx context.Context, req *logical.Request, 
 	vmName := data.Get("vm_name").(string)
 	resourceID := data.Get("resource_id").(string)
 
+	if subscriptionID != "" && !validateAzureField(guidRx, subscriptionID) {
+		return logical.ErrorResponse(fmt.Sprintf("invalid subscription id %q", subscriptionID)), nil
+	}
+	if resourceGroupName != "" && !validateAzureField(rgRx, resourceGroupName) {
+		return logical.ErrorResponse(fmt.Sprintf("invalid resource group name %q", resourceGroupName)), nil
+	}
+	if vmssName != "" && !validateAzureField(nameRx, vmssName) {
+		return logical.ErrorResponse(fmt.Sprintf("invalid vmss_name %q", vmssName)), nil
+	}
+	if vmName != "" && !validateAzureField(nameRx, vmName) {
+		return logical.ErrorResponse(fmt.Sprintf("invalid vm name %q", vmName)), nil
+	}
+
 	config, err := b.config(ctx, req.Storage)
 	if err != nil {
 		return nil, fmt.Errorf("unable to retrieve backend configuration: %w", err)